+++ ICD.Profound.ConnectPRO/Themes/UserInterface/IPresenters/Common/Settings/About/ISettingsPluginsPresenter.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Profound.ConnectPRO.SettingsTree.About;
+using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common.Settings.About;
+
+namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.Common.Settings.About
+{
+	public interface ISettingsPluginsPresenter : ISettingsNodeBasePresenter<ISettingsPluginsView, PluginsSettingsLeaf>
+	{
+	}
+}