+++ ICD.Profound.ConnectPRO/Themes/UserInterface/IViews/Common/Settings/About/ISettingsPluginsView.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+using ICD.Connect.UI.Mvp.Views;
+
+namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common.Settings.About
+{
+	public interface ISettingsPluginsView : IUiView
+	{
+		/// <summary>
+		/// Returns child views for list items.
+		/// </summary>
+		/// <param name="factory"></param>
+		/// <param name="count"></param>
+		/// <returns></returns>
+		IEnumerable<IReferencedSettingsPluginsView> GetChildComponentViews(IViewFactory factory, ushort count);
+	}
+}