+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Views/Common/Settings/About/SettingsPluginsView.generated.cs
@@ -1,0 +1,45 @@
+﻿using System.Collections.Generic;
+using ICD.Connect.Panels;
+using ICD.Connect.Panels.Devices;
+using ICD.Connect.UI.Controls;
+using ICD.Connect.UI.Controls.Lists;
+using ICD.Connect.UI.Controls.Pages;
+
+namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Views.Common.Settings.About
+{
+	public sealed partial class SettingsPluginsView
+	{
+		private VtProSubpage m_Subpage;
+		private VtProSubpageReferenceList m_PluginList;
+
+		/// <summary>
+		/// Instantiates the view controls.
+		/// </summary>
+		/// <param name="panel"></param>
+		/// <param name="parent"></param>
+		/// <param name="index"></param>
+		protected override void InstantiateControls(ISigInputOutput panel, IVtProParent parent, ushort index)
+		{
+			m_Subpage = new VtProSubpage(panel, parent, index)
+			{
+				DigitalVisibilityJoin = 146
+			};
+
+			m_PluginList = new VtProSubpageReferenceList(28, panel as IPanelDevice, m_Subpage)
+			{
+				MaxSize = 100,
+				SerialJoinIncrement = 3
+			};
+		}
+
+		/// <summary>
+		/// Gets the child controls.
+		/// </summary>
+		/// <returns></returns>
+		protected override IEnumerable<IVtProControl> GetChildren()
+		{
+			yield return m_Subpage;
+			yield return m_PluginList;
+		}
+	}
+}